:recycle: refact: header tsx전환 및 스타일 분리
